Refactory of the dependencies. Added support and tests for java.time classes in scala.js
