+++ play-json/js/src/main/scala/PlatformWrites.scala
@@ -4,6 +4,6 @@
 
 package play.api.libs.json
 
-trait EnvWrites {
+trait PlatformWrites {
   // No specific writer
 }
